Don't remove reviewers after moving from CC
The UX for moving an account from the CC reviewer state to REVIEWER
requires the user to first mark the account for pending removal from
the CC field. This queues the account to be moved to the REMOVED state
immediately after the review is posted.

Sending the reply causes the backend to move such accounts to their new
reviewer state as the user intended. However, we were immediately
deleting them from the reviewer set after posting the review. There was
no accommodation for removed accounts that are readded.

This change keeps track of accounts moving between states. They continue
to be queued for deletion when the user removes them, but any removed
account that gets readded by posting the review is now skipped during
the followup purge.

Bug: Issue 5323
Change-Id: I5918d601314504a2ca072a7f36d76bc62221803b

Former-commit-id: 64451c001a10be656566a58f6b51c7c559c8a1a4