BatchUpdate: Factor out class for repo operations
At the moment, this is basically a simple wrapper around the
repo/revwalk/inserter/commands used internally in BatchUpdate. A later
goal is to stop exposing the Repository instance directly from the
RepoView, and limit the types of repo operations available to
BatchUpdateOp implementations.

For the moment, RepoView is not documented, as its interface will change
later in this series.

Change-Id: I17a77a1f93f18fe4882d342650d0f3e0beead8e7

Former-commit-id: ca357625929383e30b976b69e0f2386dc1e015d5