Add handling for arbitrary CCs to the account-list
If allowAnyInput is enabled for the account-list, receive the string
value from the autocomplete, validate it as an email address, and add it
to the accounts array so it shows up in the account-list. If the string
is not a valid email address, reset the text in the autocomplete and
toast the user.

Also adds a function setText to gr-account-entry.

Feature: Issue 5832
Change-Id: Ia3cbc3142fb4b54fd37e5c559fddb9f425fdf948

Former-commit-id: 8cbccea11b29bfa380f2596f175cb0fe5287e478