PolyGerrit: Bug fix for iOS 10.3 and mac 10.12.4
https://bugs.webkit.org/show_bug.cgi?id=169970

There seems to be a bug in iOS 10.3 that if the header is undefined
everything fails. That's fixed in safari 10.2 which is not in safari 10.1
on iOS 10.3.

We can revert this when the fix makes it into iOS 10.3.

Bug: Issue 5715
Change-Id: I7c027b5bf90c67d15541188120b202aa965cf1d5
(cherry picked from commit 6f5ef128566ea4986bdeb70a5ba8243e58c2eaae [formerly 37140336228faaa6ff13892954bbbd425fed8a08])

Former-commit-id: 819343f9c91bd26c04be8d8362717a8d6a41d575