ExternalIdsUpdate: Add methods to replace/delete IDs across accounts
The current methods to replace and delete external IDs check that they
belong to the expected account before the write operation is done. This
is a precaution to not touch external IDs of other accounts. However
some custom extensions (e.g. plugins) want to update external IDs across
accounts in a single transaction. To support this add new methods to
ExternalIdsUpdate that replace/delete external IDs without checking to
which account they belong. These methods create a single commit in
NoteDb that do the updates.

Change-Id: I1ede49a1705eaabafc395e2f6f5c17d166a92815
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: f4391cf3aca51ea1bc8b5ed7d5b685b1d9a0ebc2