ChangeApiImpl: Don't catch RestApiException when editing hashtag
If a RestApiException (or, importantly, any subclass of RestApiException)
is raised when performing the hashtag edit, it is caught and rethrown as
RestApiException with a generic "Cannot post hashtags" message.

Any specific exception type, and its message, is then lost.

Don't catch RestApiException. Instead let it propagate up to the caller.

Also add tests to confirm the behavior when adding hashtags with and
without the "edit hashtags" permission.

Change-Id: Ib708d871462972b3f0644351d062de08ef1c9bee

Former-commit-id: 483baef151bdf7ea96afae7cddbd9dc249e72aec