ChangeApiImpl: Don't catch RestApiException on rebase
RestApiException (and any subclass of it) is caught and rethrown as
a RestApiException with a generic message. This causes the actual
exception type and message to be lost.

Don't catch RestApiException; allow the original to propagate.

Also add tests in ChangeIT to verify behavior when rebasing with and
without the rebase permission.

Change-Id: Ia303dcf0b7158aafa82adfaeeaabd76a8a980b15

Former-commit-id: 52e6bc98b146cc5fbe0dd2af3c2b2b18bbb79e37