Remove bouncycastle jar files from lib folder during init
Bouncycastle is now shipped in the war file, and does not need to be
separately downloaded. However, when upgrading from an older version,
the previously downloaded jar files remain in the site's lib folder.

The LibraryDownloader already has functionality to remove (actually
rename) the stale libraries, but that only works when those libraries
are handled as downloads.

Factor the stale library removal functionality out to a separate class
and use that to remove the bouncycastle jar files during init.

Change-Id: I048250026cb1040b75c9b34a4aaf852e5e3384c6

Former-commit-id: 50c9b5bdeabb53552638a77b42590d2b7f171201