Fix the wrong order of RevWalk.isMergedInto in RebaseSorter
RevWalk.isMergedInto takes the base as first argument and the tip as
the second argument to check if the base is merged into the tip. The
test didn't detect it because in the test the base is the same as the
tip.

Also add a regression test for this bug.

Change-Id: I43742aa4bd75ea64e729c1d3ecf40369002c7e55

Former-commit-id: 15e9a266b3a9cb06f8ed131482409b7fc4c73278