Don't leak ObjectReaders created from ObjectInserters
We had been pretty lazy about calling new RevWalk(inserter.newReader()).
I can't say with certainty for any individual case, but there are
at least two possible loose justifications for this:

1. ObjectInserter#newReader() feels like it shouldn't allocate any new
   resources (other than GCable buffers).
2. It seems like the RevWalk might close the reader.

Both of these are bad arguments. For (1), WindowCursor acquires an
inflater from the inflater cache, so failing to release the inflater
results in more allocations. For (2), RevWalk does not close the reader
when using this constructor. While perhaps non-obvious, this behavior is
clearly documented and stable.

Change-Id: I712973ecfcca439245722c777abde129dbe3773d

Former-commit-id: d3239cced271000557cd5410dd0976277f9c8a55