Make sure path in switch account URL is up to date
This URL is specified as an entry in the links property. Recently we
changed links into a computed property, so it could depend on the _path
property. We forgot to list _path as one of the dependencies.

Now we derive links from both switchAccountUrl and _path. The former is
updated by loading the server config, and the latter is updated by
listening to location changes.

Bug: Issue 5921
Change-Id: Id1b6eb46308de8ce45cbaaaf96240b2a5090c000

Former-commit-id: 18664b76e1419d0b4279fc86942716f392e2132c