Use start of comment range for email parser
When sending emails, we don't use the line number, but the start of the
range comment to link back to the comment. The email parser was
exclusively using lineNbr which caused a comment to be identified as a
file comment instead of a reply to an inline comment.

This changes fixes the problem.

Bug: Issue 5794
Change-Id: Ib7690f619793d0db703eecd5cfe2b4f2ad6c9ede

Former-commit-id: cdeb08b3ea8dbb027f1fff617f5921321be5a339