StalenessChecker: Change is stale if primary storage doesn't match
After running PrimaryStorageMigrator on a change, the note_db_state
field changes, but the change is not automatically reindexed. The idea
was that the change should be detected as stale--after all, its
row_version and note_db_state changed. Unfortunately, the implementation
of StalenessChecker doesn't actually flag a change as stale in this
case, because it ignores row_version entirely when the change is NoteDb
primary.

Worse, the now-stale value in the index still doesn't include the change
meta ref in the ref_state field, since that ref isn't recorded when
ReviewDb is primary, so we aren't able to check for staleness using the
NoteDb state either.

Break this stalemate by considering a change as stale if its
PrimaryStorage in the index differs between the version of the Change
located in primary storage and the version in the index.

Change-Id: I3c5d9981bf6e8a465509d94cac13c66e8368bdbe

Former-commit-id: 98cd9ea59b2f8f93d229976184b556d19c3a6585