Never autoRebuild when ReviewDb is disabled for changes
Changes can't be rebuilt when their primary storage is NoteDb.
Generally, when ReviewDb is disabled for changes, we assume that it is
not possible to encounter a change with PrimaryStorage == REVIEW_DB, so
it seemed like checking the PrimaryStorage when setting the autoRebuild
bit was sufficient.

However, in the corner case described in I3c5d9981, it is possible to
encounter a REVIEW_DB change via deserializing a stale Change proto from
the index, even though the actual primary storage of every change is
NOTE_DB. Trying to auto-rebuild in this case will result in an error,
as ChangeRebuilder rereads the change from primary storage and will
refuse to rebuild a NoteDb-primary change.

Ensure we never hit the auto-rebuilding path when ReviewDb is disabled
by explicitly checking the state of the NotesMigration, rather than
depending only on the PrimaryStorage.

Change-Id: Iae54fbe94f85e57dee6a2c9848b89e0697ff88e6

Former-commit-id: 8f9e11957a1c1ba3fd39e7a5d58196d0d9972116