Switch GitHub mirror URL to the most up-to-date
There are two mirrors in GitHub, one that is updated very frequently
(https://github.com/GerritCodeReview) and contains all the Gerrit repos
including plugins, and another that is updated rarely (https://github.com/gerrit-review)
and doesn't contain all plugins.

Switch the URL to the most up-to-date and complete repo on GitHub.

Change-Id: I7037c52e58919dc237eaa39bd23b2cf14ab5ada9

Former-commit-id: fff21acc042f965aa2d1144330f675865b1c4020