Hide assignee by default in Gerrit changes table
Keep backward-compatible behaviour for users coming
from Gerrit < 2.14 and not interested in the assignee
workflow.

The assignee feature will still be accessible through a second
configuration setting (showAssigneeSuggestOracle) that controls
the ability to suggest searching of the changes by assignee.

Typically people and organisations that want to use the
feature massively, will then enable it in the gerrit.config
to be more prevalent on the changes table.

When both showAssignee and showAssigneeSuggestOracle are
set to false, the feature is not accessible anymore because
you cannot neither see nor search for changes by assignee.

Change-Id: Ie826c126b5508c9013ee4f802f6bcfaff95cef76

Former-commit-id: 50057ff887df2fe7c9607bf5560554a50535189d