Fix dropdown tap on touch device
Previously, there was an issue where click events did not fire before
tap handlers closed a dropdown. If the dropdown was closed first, the
dropdown is gone and something random becomes the event target.

This change adds a short setTimeOut to ensure that the link is the click
target prior to closing the dropdown.

Bug: Issue 5932
Change-Id: I218893fd479eeb68361dd2c94d919740a8f542f2

Former-commit-id: 03049078177a621e3967b07b779396a77aa59471