PluginLoader#load: Fix NPE when no plugins used on a site
This fixes a NPE that was introduced in I19e564891, when no plugins
are used on gerrit site.

As the direct consequence of this error, inline edit feature is broken,
because host page load sequence is interrupted and edit preferences
cannot be loaded and thus inline editor cannot be initialized.

Bug: Issue 5953
Bug: Issue 5929
Change-Id: I84b07aeb0283cc10d2dffa9ca04769b53f8e8b01

Former-commit-id: 721ac978a986de0f322528cd6c6f70389aedf1f7