Add modifier check to file list keyboard shortcuts
The chrome browser has a keyboard shortcut for jumping to the bottom of
the page. The file list down arrow intercepted that shortcut so it did
not function as expected. This change adds a modifier check for up and
down arrows so if a modifier is used with one of the keys, it won't
prevent default browser behavior.

Bug: Issue 5949
Change-Id: Ie7139436b74008afe2ad829728000b14cccde0db

Former-commit-id: 99e8d7bd95ce70d61bf7d841607608d561e188b4