GET files?reviewed: Don't fire N+1 selects for N patch sets
If current patch set doesn't have any reviewed bit set on any files set
by current user, the algorithm iterates over all patch sets in reversed
order, fires dedicated select statement for each patch set and copies
the reviewed bits from the latest patch set found, where reviewed bits
were set. There is off by one mistake in the current implementation, so
that the select for the current patch set is fired twice.

In the worst case the current implementation also fires N+1 select
statements for N patch set. Especially on big gerrit sites, we have seen
changes with three digits patch set number, so that we eventualy spawn
hundreds of select statements with the current approach to load only one
single change.

This is a known anti pattern in SQL DML domain, to span dynamic number
of select statements, that depends on the current input (change).

To rectify, express the desired query with SQL. For a given user, change
and specific patch set (that is not necessarily the current patch set of
the change), find a patch set with max patch set number, that is smaller
or equals to the given patch set, with reviewed bit set.

Bug: Issue 5907
Change-Id: I09d345f6a396d1c9ceea4a2135f539ea073f59cd

Former-commit-id: f3cdd45e9e3bc2361e87ae2542079331e20c93cd