PluginLoader: Remove stale plugin files from /tmp on startup
When plugins are loaded, they are copied to /tmp and run from that
location. A CleanupHandler is created, which then deletes the file
from /tmp on server shutdown.

However, if the server is not shutdown gracefully, the cleanup does
not get run and the file remains in the /tmp folder. If this happens
frequently, the number of stale files in /tmp increases over time.

On startup, scan the /tmp folder for plugin files and remove them.

Bug: Issue 5965
Change-Id: Ic58c84cabee747981e8149180c597b7bc7440e3a

Former-commit-id: 4bb459957c3d7d2b03d0fefff2ee7e8cb79b0c8c