Conditionally disable assignee editing
Before, the assignee field was mutable if the user was logged in. This
was an inadequate representation of whether or not editing the assignee
was a valid action.

This change uses `actions.assignee.enabled` as well as the mutable flag
to determine whether or not the assignee field is editable.

Bug: Issue 5988
Change-Id: I57bdd72e054fe6e75481636f7b191a7e4dd93504

Former-commit-id: 486b0aa8391229352945405ef62c3ee95cb072f1