Disallow deletion of user branches
When accounts are migrated to NoteDb the existence of a user branch is
mandatory for an account. Deleting the user branch would mean deleting
the account but we don't support account deletion. Hence disallow
deletion of user branches even if the user has the DELETE permission.
Only allow the deletion of a user branch if the calling user also has
the ACCESS_DATABASE capability. With this capability an account entry
can be deleted in ReviewDb, hence it should allow to do the same in
NoteDb.

Change-Id: I5e67a86b51a038483222f3e57bd1399a9f290db3
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 8b7c852b13cb454d83e574d8049561cf086e3615