Move repository registration out of AbstractDaemonTest#cloneProject
By moving out the repository registration, we can get the registered URI
which can be used to connect to the repository for both fetch and push.

This makes it possible to clone a remote repository to another remote
repository in the integration test and both of them are managed by
GitRepositoryManager.

Change-Id: Ic941ae184249bb9521da52350ae45b24cfe5f8c9

Former-commit-id: 9a3b435080fe21b30155a2b7031572b33dd2897d