Fix comment navigation in file list
Previously, the check for modifiers came before the check for shift,
so jumping to the next comment thread did not work properly.

This change checks for modifiers, so long as it is not the shift key,
and also checks for shift in the event detail and in the
keyboardEvent, as is done in the check for meta keypress in general.

Bug: Issue 6010
Change-Id: Iadfda322400d59450b62c879d55ed5a6e3f233df

Former-commit-id: 7e218ddcaf71e7722848c79522e8ab9be818b7b2