Allow disabling ReindexAfterRefUpdate
While it's a nice idea in theory to recheck the mergeable bit on all
open changes when the branch advances, or to reindex everything when
label definitions change, in practice it can be quite taxing on a live
server with thousands of open changes on an active project. Allow admins
to disable this if necessary.

Change-Id: I551a2afa9572e71e36b5344a059346eb92cb2fcf

Former-commit-id: 6115562f9577215d83f7f1dd1dc8b9cec54be69a